feat: remove the feature to hide comments from post and exec command
